Move ProjectManager and ProjectManagerDecorator on board of libmvvm_model
