+++ examples/quickrefl/core/editors/welcomeview/welcomeview.cpp
@@ -10,7 +10,7 @@
 #include "welcomeview.h"
 #include "applicationmodels.h"
 #include "openprojectwidget.h"
-#include "projectmanagerdecorator.h"
+#include <mvvm/project/projectmanagerdecorator.h>
 #include "projectmanagerinteractor.h"
 #include <mvvm/project/projectutils.h>
 #include "recentprojectwidget.h"
