+++ examples/quickrefl/tests/projectmanager.test.cpp
@@ -10,7 +10,7 @@
 #include <mvvm/interfaces/applicationmodelsinterface.h>
 #include "folderbasedtest.h"
 #include "google_test.h"
-#include "projectmanager.h"
+#include <mvvm/project/projectmanager.h>
 #include "test_utils.h"
 #include <cctype>
 #include <mvvm/model/propertyitem.h>
