+++ examples/quickrefl/tests/projectmanagerdecorator.test.cpp
@@ -7,14 +7,14 @@
 //
 // ************************************************************************** //
 
-#include <mvvm/interfaces/applicationmodelsinterface.h>
 #include "folderbasedtest.h"
 #include "google_test.h"
-#include "project_types.h"
-#include "projectmanagerdecorator.h"
 #include <cctype>
+#include <mvvm/interfaces/applicationmodelsinterface.h>
 #include <mvvm/model/propertyitem.h>
 #include <mvvm/model/sessionmodel.h>
+#include <mvvm/project/project_types.h>
+#include <mvvm/project/projectmanagerdecorator.h>
 #include <mvvm/utils/fileutils.h>
 
 namespace
