+++ source/libmvvm_model/mvvm/project/CMakeLists.txt
@@ -3,8 +3,13 @@ target_sources(mvvm_model PRIVATE
     modelhaschangedcontroller.h
     project.cpp
     project.h
+    project_types.h
     projectchangecontroller.cpp
     projectchangecontroller.h
+    projectmanager.cpp
+    projectmanager.h
+    projectmanagerdecorator.cpp
+    projectmanagerdecorator.h
     projectutils.cpp
     projectutils.h
 )
